conversion de generate-types.js en typescript + optim + retrait de la liste hard-codée
